Fixed bugs detected with warnings,  utopia, utopia_test, and utopia_bench are warning free (on MacOSX, HPC, and Daint)

+++ utopia/backend/trilinos/utopia_Tpetra_Vector.cpp
@@ -388,7 +388,7 @@ namespace utopia {
                               const Scalar &val) {
         comm_ = comm;
 
-        if (n_local == Traits<TpetraVector>::decide()) {
+        if (static_cast<std::size_t>(n_local) == Traits<TpetraVector>::decide()) {
             values(n_global, val);
             return;
         }
